rag/requirements.txt
llama-index>=0.12.0
llama-index-core>=0.12.0

llama-index-llms-dashscope
llama-index-embeddings-dashscope

llama-index-vector-stores-chroma
llama-index-readers-file
chromadb>=0.5.0

pypdf>=4.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.23
pandas>=2.0.0

python-dotenv>=1.0.0

tqdm
colorama
psutil
pycryptodome
docx2txt
